Revert "Move `TextInput` to checkout package and allow it to be used for input type=number (#4238)"
This reverts commit 15c17794e67aa85a11ed92606e115741f29e57fb.
