+++ assets/js/base/components/state-input/state-input.js
@@ -3,18 +3,14 @@
  */
 import { __ } from '@wordpress/i18n';
 import PropTypes from 'prop-types';
-import { ValidatedTextInput } from '@woocommerce/blocks-checkout';
 import { decodeEntities } from '@wordpress/html-entities';
 import { useCallback, useMemo } from '@wordpress/element';
 import classnames from 'classnames';
-import {
-	useValidationContext,
-	ValidationInputError,
-} from '@woocommerce/base-context';
 
 /**
  * Internal dependencies
  */
+import { ValidatedTextInput } from '../text-input';
 import { ValidatedSelect } from '../select';
 import './style.scss';
 
@@ -29,22 +25,6 @@ const StateInput = ( {
 	value = '',
 	required = false,
 } ) => {
-	const {
-		getValidationError,
-		getValidationErrorId,
-		setValidationErrors,
-		clearValidationError,
-		hideValidationError,
-	} = useValidationContext();
-
-	const textInputValidationFunctions = {
-		getValidationError,
-		getValidationErrorId,
-		setValidationErrors,
-		clearValidationError,
-		hideValidationError,
-	};
-
 	const countryStates = states[ country ];
 	const options = useMemo(
 		() =>
@@ -128,8 +108,6 @@ const StateInput = ( {
 			autoComplete={ autoComplete }
 			value={ value }
 			required={ required }
-			inputErrorComponent={ ValidationInputError }
-			{ ...textInputValidationFunctions }
 		/>
 	);
 };
