+++ assets/js/base/components/text-input/style.scss
@@ -39,7 +39,6 @@
 
 	input[type="tel"],
 	input[type="url"],
-	input[type="number"],
 	input[type="text"],
 	input[type="email"] {
 		@include font-size(regular);
@@ -75,7 +74,6 @@
 
 	&.is-active input[type="tel"],
 	&.is-active input[type="url"],
-	&.is-active input[type="number"],
 	&.is-active input[type="text"],
 	&.is-active input[type="email"] {
 		padding: em($gap-large) 0 em($gap-smallest) $gap;
