+++ assets/js/base/components/text-input/text-input.tsx
@@ -4,12 +4,12 @@
 import { forwardRef, InputHTMLAttributes } from 'react';
 import classnames from 'classnames';
 import { useState } from '@wordpress/element';
+import { Label } from '@woocommerce/blocks-checkout';
 
 /**
  * Internal dependencies
  */
 import './style.scss';
-import Label from '../label';
 
 interface TextInputProps
 	extends Omit<
@@ -49,21 +49,18 @@ const TextInput = forwardRef< HTMLInputElement, TextInputProps >(
 				/* Do nothing */
 			},
 			feedback,
-			...rest
 		},
 		ref
 	) => {
 		const [ isActive, setIsActive ] = useState( false );
+
 		return (
 			<div
 				className={ classnames(
 					'wc-block-components-text-input',
 					className,
 					{
-						'is-active':
-							isActive ||
-							value ||
-							( value === 0 && type === 'number' ),
+						'is-active': isActive || value,
 					}
 				) }
 			>
@@ -90,7 +87,6 @@ const TextInput = forwardRef< HTMLInputElement, TextInputProps >(
 							: ariaDescribedBy
 					}
 					required={ required }
-					{ ...rest }
 				/>
 				<Label
 					label={ label }
