+++ assets/js/blocks/cart-checkout/checkout/form/contact-fields-step.js
@@ -3,12 +3,8 @@
  */
 import { __ } from '@wordpress/i18n';
 import { FormStep } from '@woocommerce/base-components/cart-checkout';
-import { ValidatedTextInput } from '@woocommerce/blocks-checkout';
-import {
-	useCheckoutContext,
-	useValidationContext,
-	ValidationInputError,
-} from '@woocommerce/base-context';
+import { ValidatedTextInput } from '@woocommerce/base-components/text-input';
+import { useCheckoutContext } from '@woocommerce/base-context';
 import { getSetting } from '@woocommerce/settings';
 import CheckboxControl from '@woocommerce/base-components/checkbox-control';
 
@@ -27,21 +23,6 @@ const ContactFieldsStep = ( {
 		shouldCreateAccount,
 		setShouldCreateAccount,
 	} = useCheckoutContext();
-	const {
-		getValidationError,
-		getValidationErrorId,
-		setValidationErrors,
-		clearValidationError,
-		hideValidationError,
-	} = useValidationContext();
-
-	const textInputValidationFunctions = {
-		getValidationError,
-		getValidationErrorId,
-		setValidationErrors,
-		clearValidationError,
-		hideValidationError,
-	};
 
 	const createAccountUI = ! customerId &&
 		allowCreateAccount &&
@@ -80,8 +61,6 @@ const ContactFieldsStep = ( {
 				autoComplete="email"
 				onChange={ onChangeEmail }
 				required={ true }
-				inputErrorComponent={ ValidationInputError }
-				{ ...textInputValidationFunctions }
 			/>
 			{ createAccountUI }
 		</FormStep>
